feat(amount): move util functions to lib, add function to generate formatted amount in string
